+++ changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to Lock - Laravel will be documented in this file. This file follows the *[Keep a CHANGELOG](http://keepachangelog.com/)* standards.
 
+## Unreleased
+
+### Fixed
+
+- Fixed a bug where the guest caller had an incorrect id set
+
 ## 0.1.1 - 2014-12-17
 
 ### Fixed