+++ changelog.md
@@ -2,7 +2,7 @@
 
 All notable changes to Lock - Laravel will be documented in this file. This file follows the *[Keep a CHANGELOG](http://keepachangelog.com/)* standards.
 
-## Unreleased
+## 0.1.2 - 2015-01-13
 
 ### Changed